LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/transition-base-response-002-expected.txt



FAIL Transitioning font-size on root affects rem units assert_equals: expected "15px" but got "16.5px"
